# swagger-web/analysis-swg/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Autenticación, Reportes y Análisis",
    "description": "API para iniciar sesión, obtener reportes de análisis y realizar análisis de dispositivos",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Autenticación de usuario y generación de tokens JWT",
        "description": "Este endpoint permite a un usuario autenticarse mediante un nombre de usuario y una contraseña. Si las credenciales son válidas, se generan y devuelven un access token y un refresh token.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string",
                  "example": "admin"
                },
                "password": {
                  "type": "string",
                  "example": "MySecurePass123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticación exitosa. Se devuelve el access token y refresh token.",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                },
                "refresh_token": {
                  "type": "string",
                  "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas (usuario o contraseña incorrectos).",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad username or password"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Generic Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Renovar el access token usando el refresh token",
        "description": "Este endpoint permite renovar el access token usando el refresh token proporcionado. El usuario debe estar autenticado mediante JWT.",
        "responses": {
          "200": {
            "description": "Se genera un nuevo access token.",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido o expirado.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid refresh token"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "El token JWT debe ser proporcionado en la cabecera de la solicitud."
    }
  }
}